refactor(box): split components into separate files
